move notebook list from MainActivity to HomeFragment and add other basic fragments for home screen
