+++ app/src/main/java/com/github/bytesculptor07/quillo/SharednotebooksFragment.java
@@ -1,0 +1,23 @@
+package com.github.bytesculptor07.quillo;
+
+import android.os.Bundle;
+
+import androidx.fragment.app.Fragment;
+
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+public class SharednotebooksFragment extends Fragment {
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_sharednotebooks, container, false);
+    }
+}
