+++ app/src/main/res/layout/activity_main.xml
@@ -10,70 +10,11 @@
      android:id="@+id/parentLayout">
 
     
-    <LinearLayout
-         android:layout_height="match_parent"
-         android:layout_width="match_parent"
-         android:orientation="vertical"
-         android:layout_margin="20dp"
-         android:layout_weight="3">
-        
-        <RelativeLayout
-             android:layout_height="wrap_content"
-             android:layout_width="match_parent">
-        
-            <LinearLayout
-                android:layout_height="wrap_content"
-                android:layout_width="wrap_content"
-                android:id="@+id/buttonPanel"
-                android:layout_centerInParent="true"
-                android:layout_alignParentLeft="true">
-
-                <Button
-                    android:layout_height="wrap_content"
-                    android:layout_width="wrap_content"
-                    android:id="@+id/create"
-                    android:layout_margin="10dp"
-                    android:textColor="@color/white"
-                    android:drawableLeft="@drawable/icon_add"
-                    android:text="@string/button_create"/>
-
-                <Button
-                    android:layout_height="wrap_content"
-                    android:layout_width="wrap_content"
-                    android:layout_margin="10dp"
-                    style="@style/Widget.Material3.Button.OutlinedButton"
-                    android:text="@string/button_sort_order"/>
-        
-            </LinearLayout>
-            
-            <ImageView
-                 android:layout_height="50dp"
-                 android:layout_width="50dp"
-                 android:layout_margin="10dp"
-                 android:layout_centerInParent="true"
-                 android:layout_alignParentRight="true"
-                 android:src="@drawable/icon_user" />
-            
-        </RelativeLayout>
-        
-        <View
-            android:layout_height="1dp"
-            android:layout_width="match_parent"
-            android:layout_margin="10dp"
-            android:background="#aaaaaa"/>
-
-        <ScrollView
-            android:layout_height="match_parent"
-            android:layout_width="match_parent" >
-        
-            <LinearLayout
-                android:layout_height="match_parent"
-                android:layout_width="match_parent"
-                android:orientation="vertical"
-                android:id="@+id/folderList" />
-        </ScrollView>
-
-    </LinearLayout>
+    <FrameLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:id="@+id/fragmentContainer">
+    </FrameLayout>
     
     <com.google.android.material.bottomnavigation.BottomNavigationView
         android:id="@+id/bottom_navigation"
